refactor(nimbus): extract useFocusInputOnFieldClick and share across inputs
Extract the click-to-focus behavior into a shared useFocusInputOnFieldClick
hook and migrate both SearchInput and TextInput onto it. The hook uses the
scoped mousedown + preventDefault approach (no focus flicker, ignores
interactive children like the clear button), replacing TextInput's prior
global document click listener.

Also trims the inline comments on SearchInput and its recipe, and extends the
changeset to cover TextInput for easier triage.
