+++ packages/nimbus/src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from "./use-color-mode";
 export * from "./use-color-mode-value";
 export * from "./use-localized-string-formatter";
 export * from "./use-drag-and-drop";
+export * from "./use-focus-input-on-field-click";
