+++ packages/nimbus/src/hooks/use-focus-input-on-field-click/index.ts
@@ -1,0 +1,1 @@
+export { useFocusInputOnFieldClick } from "./use-focus-input-on-field-click.ts";
